requirement.txt
agno
python-dotenv
faiss-cpu
sentence-transformers
PyPDF2
numpy